moved images into their own directory
