+++ client/src/components/Navigation.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import logo from './logo.png';
+import logo from '../images/logo.png';
 
 const Navigation = () => {
   return (
