+++ client/src/components/OneCity.jsx
@@ -1,7 +1,5 @@
 import React from 'react';
-import { aqiToCig, hazardLevels } from '../constants/constants';
-import { solutionsObj } from "../constants/constants"
-
+import { aqiToCig, hazardLevels, solutionsObj } from '../constants/constants.js';
 
   const OneCity = ({ location }) => {
     const {
