+++ client/src/components/PollutionStats.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import '../styling/App.css';
 import Card from './Card';
+
 const PollutionStats = ({ stats, cityUrl, remove, weather }) => {
   const timeStamp = stats.ts.slice(0, 10);
   return (
